Move server to cloud (#9)
