+++ ROKTION-server/src/schemas/doc.ts
@@ -1,15 +1,15 @@
 import { Document, Schema, model } from "mongoose";
 
-export interface Docs extends Document {
+export interface Doc extends Document {
     pageTitle: string;
     content: string;
     linkedFiles: Array<[string, string]>;
 }
 
-const docsSchema = new Schema({
+const docSchema = new Schema({
     pageTitle: String,
     content: String,
     linkedFiles: Array
 });
 
-export const DocsModel = model<Docs>('Docs', docsSchema);
+export const DocModel = model<Doc>('Doc', docSchema);
